pyproject.toml
[project]
name = "threat_modeling"
version = "0.1.0"
description = "Threat modeling using CrewAI"
authors = [
    { name = "Charles Green", email = "hello@charles.green" },
]
requires-python = ">=3.10,<=3.13"

dependencies = [
    "crewai==0.119.0",
    "crewai-tools==0.44.0",
    "langfuse==2.60.3",
    "pymupdf==1.23.6",
    "pillow==10.2.0",
    "pytesseract==0.3.10",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3"
]

[project.scripts]
threat_modeling = "threat_modeling.main:run"
run_crew = "threat_modeling.main:run"
train = "threat_modeling.main:train"
replay = "threat_modeling.main:replay"
test = "threat_modeling.main:test"

[build-system]
requires = [
    "hatchling"
]
build-backend = "hatchling.build"
